replace custom_conf.py by conf.py
